+++ core/src/mast/mod.rs
@@ -12,6 +12,8 @@ use core::{
 use serde::{Deserialize, Serialize};
 
 mod node;
+#[cfg(any(test, feature = "arbitrary"))]
+pub use node::arbitrary;
 pub use node::{
     BasicBlockNode, CallNode, DecoratedOpLink, DecoratorOpLinkIterator, DynNode, ExternalNode,
     JoinNode, LoopNode, MastNode, MastNodeErrorContext, MastNodeExt, OP_BATCH_SIZE, OP_GROUP_SIZE,
@@ -48,6 +50,10 @@ mod tests;
 /// can be built from a [`MastForest`] to specify an entrypoint.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(
+    all(feature = "serde", feature = "arbitrary", test),
+    miden_serde_test_macros::serde_test
+)]
 pub struct MastForest {
     /// All of the nodes local to the trees comprising the MAST forest.
     nodes: Vec<MastNode>,